requirements.txt
psycopg2-binary==2.9.9
Flask==2.2.3
Flask-Cors==4.0.0
openai==0.28.1
python-dotenv==1.0.0